feat(release): version 2.2.0 — functional data plane, persistence & durability
Turn cynamoDB into a working local DynamoDB replacement. The prior release built but the data path was non-operational (stubbed storage engines, an HTTP layer that never routed requests, and a server that was never started).

- Implement the HTTP/JSON data plane: CreateTable/DescribeTable/ListTables, PutItem/GetItem/DeleteItem, Scan (paginated), Query (partition-key EQ, sorted); start the server in main.
- Implement the storage engines (MemoryEngine + LsmEngine) with an order-preserving key codec; fix tombstone reads and per-table isolation.
- Persistence: WAL replay, SSTable reload from manifest, atomic table-catalog persistence; real compaction (merge + tombstone purge, bounded file count); fsync-per-write WAL durability (survives kill -9 and power loss).
- Fix a metrics use-after-free, an LSM shutdown deadlock, and a compaction busy-spin.
- Add rigorous unit + live HTTP integration tests (CRUD across restarts, SIGKILL crash recovery, concurrent load), green under ASan/UBSan.

See CHANGELOG.md and ITEMS_TO_FIX.md for details.

+++ src/streams/manager.cpp
@@ -13,7 +13,9 @@ std::string generate_event_id() {
     static thread_local std::mt19937 rng{std::random_device{}()};
     std::uniform_int_distribution<uint64_t> dist;
     char buf[33];
-    snprintf(buf, sizeof(buf), "%016llx%016llx", dist(rng), dist(rng));
+    snprintf(buf, sizeof(buf), "%016llx%016llx",
+             static_cast<unsigned long long>(dist(rng)),
+             static_cast<unsigned long long>(dist(rng)));
     return std::string(buf);
 }
